+++ tests/Unit/Requests/PutRequestTest.php
@@ -28,7 +28,7 @@ use PHPUnit\Framework\TestCase;
 
 final class PutRequestTest extends TestCase
 {
-	public function testRequestMethodIsPut()
+	public function testRequestMethodIsPut() : void
 	{
 		$request = new PutRequest( '/path/to/script.php', 'Unit-Test' );