+++ tests/bootstrap.php
@@ -21,4 +21,4 @@
  * SOFTWARE.
  */
 
-require(__DIR__ . '/../vendor/autoload.php');
+require __DIR__ . '/../vendor/autoload.php';